requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.0.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
firebase-admin==5.4.0
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4
idna==3.4
msgpack==1.0.4
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
